sprint3 mameno hashmap
