--- Framework/WEB-INF/classes/Mapping.java
@@ -1,19 +1,0 @@
-package etu1929.framework;
-
-public class Mapping {
-    String className;
-    String method;
-    
-    public void setClassName(String className) {
-        this.className = className;
-    }
-    public void setMethod(String method) {
-        this.method = method;
-    }
-    public String getClassName() {
-        return className;
-    }
-    public String getMethod() {
-        return method;
-    }
-}
